Cleaned up tests. Vertically centered processing pages.
