+++ frontend/app/preview/processing/page.tsx
@@ -35,21 +35,23 @@ export default function PreviewProcessing() {
 	}, [router]);
 
 	return (
-		<div className="flex flex-col items-center justify-center h-screen">
-			<h1 className="text-3xl font-semibold mb-4">We are processing your job!</h1>
-			<Spinner />
-			<div className="mb-3" />
-			<Progress
-				aria-label="Downloading..."
-				className="max-w-md"
-				showValueLabel={true}
-				size="md"
-				value={progress}
-			/>
-			<div className="mb-3" />
-			<p className="text-lg mt-4">
-				Your job is being processed. You will be redirected to the download page once it&#39;s ready.
-			</p>
+		<div className="flex flex-col items-center justify-center h-full">
+			<div className="flex flex-col items-center justify-center">
+				<h1 className="text-3xl font-semibold mb-4">We are processing your job!</h1>
+				<Spinner />
+				<div className="mb-3" />
+				<Progress
+					aria-label="Downloading..."
+					className="max-w-md"
+					showValueLabel={true}
+					size="md"
+					value={progress}
+				/>
+				<div className="mb-3" />
+				<p className="text-lg mt-4">
+					Your job is being processed. You will be redirected to the download page once it&#39;s ready.
+				</p>
+			</div>
 		</div>
 	);
 }
