+++ frontend/package.json
@@ -26,6 +26,7 @@
 		"intl-messageformat": "^10.5.0",
 		"next": "^15.1.6",
 		"next-themes": "^0.4.4",
+		"playwright": "^1.51.1",
 		"posthog-js": "^1.231.2",
 		"react": "18.3.1",
 		"react-dom": "18.3.1",
