+++ frontend/tests/login.test.ts
@@ -1,8 +1,10 @@
 import { test, expect } from "@playwright/test";
 
 test.describe("Login Tests", () => {
+	const HOMEPAGE_URL = "http://localhost:3000";
+
 	test.beforeEach(async ({ page }) => {
-		await page.goto("http://localhost:3000");
+		await page.goto(HOMEPAGE_URL);
 	});
 
 	test("should display login button in homepage navbar and navigate to Google auth", async ({ page }) => {
